requirements.txt
academictorrents==2.3.3
albumentations==1.4.14
einops==0.8.0
imageio==2.35.1
more_itertools==10.5.0
numpy==2.1.1
omegaconf==2.3.0
Pillow==10.4.0
pudb==2024.1.2
pytorch_lightning==1.9.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scipy==1.14.1
setuptools==59.6.0
skimage==0.0
streamlit==1.38.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
